uda/meme-gen/meme-generator-done/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docx==0.2.4
Flask==1.1.1
graphviz==0.13.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
meme==2.0.0
numpy==1.17.4
pandas==0.25.3
pdftotext==2.1.2
Pillow==6.2.1
pyception==0.1.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
